Seperated BoardsAddSection section from ViewBoard page
